refactor: refactoring update to reuse _get_latest_md
this refactor will allow reusing _get_latest_md by delegations
